Refactor jsch usage to blueocean-git-pipeline
